api refactor - moved remaining api traits to api module
