+++ contracts/examples/lottery-egld/src/lottery_info.rs
@@ -1,4 +1,5 @@
-use elrond_wasm::{Address, BigUintApi, Vec};
+use elrond_wasm::api::BigUintApi;
+use elrond_wasm::{Address, Vec};
 
 derive_imports!();
