+++ contracts/examples/lottery-erc20/src/lottery_info.rs
@@ -1,4 +1,5 @@
-use elrond_wasm::{Address, BigUintApi, Vec};
+use elrond_wasm::api::BigUintApi;
+use elrond_wasm::{Address, Vec};
 derive_imports!();
 
 #[derive(NestedEncode, NestedDecode, TopEncode, TopDecode, TypeAbi)]
