+++ elrond-wasm-debug/src/abi_json/mod.rs
@@ -8,7 +8,7 @@ pub use type_abi_json::*;
 
 /// Function provided for convenience.
 /// Yields the ABI JSON of a contract as string.
-pub fn contract_abi<C: elrond_wasm::ContractWithAbi>(contract: &C) -> String {
+pub fn contract_abi<C: elrond_wasm::api::ContractWithAbi>(contract: &C) -> String {
 	let abi = contract.abi(true);
 	serialize_abi_to_json(&abi)
 }
