+++ elrond-wasm-debug/src/api/endpoint_arg_api_mock.rs
@@ -1,9 +1,9 @@
 use crate::{TxContext, TxPanic};
 use alloc::vec::Vec;
+use elrond_wasm::api::ContractHookApi;
 use elrond_wasm::api::{EndpointArgumentApi, ErrorApi};
 use elrond_wasm::err_msg;
 use elrond_wasm::types::BoxedBytes;
-use elrond_wasm::ContractHookApi;
 use num_bigint::{BigInt, BigUint};
 use num_traits::cast::ToPrimitive;
