+++ elrond-wasm-debug/src/blockchain_mock.rs
@@ -6,10 +6,10 @@ use crate::contract_map::*;
 use crate::display_util::*;
 use crate::ext_mock::*;
 
+use elrond_wasm::api::BigUintApi;
+use elrond_wasm::api::CallableContract;
+use elrond_wasm::api::ContractHookApi;
 use elrond_wasm::err_msg;
-use elrond_wasm::BigUintApi;
-use elrond_wasm::CallableContract;
-use elrond_wasm::ContractHookApi;
 
 use num_bigint::{BigInt, BigUint};
 use num_traits::{cast::ToPrimitive, Zero};
