+++ elrond-wasm/src/api/mod.rs
@@ -1,3 +1,8 @@
+mod big_int_api;
+mod big_uint_api;
+mod callable_contract_api;
+mod contract_abi_api;
+mod contract_api;
 mod crypto_api;
 mod endpoint_arg_api;
 mod endpoint_finish_api;
@@ -5,6 +10,11 @@ mod error_api;
 mod log_api;
 mod storage_api;
 
+pub use big_int_api::*;
+pub use big_uint_api::*;
+pub use callable_contract_api::*;
+pub use contract_abi_api::*;
+pub use contract_api::*;
 pub use crypto_api::*;
 pub use endpoint_arg_api::*;
 pub use endpoint_finish_api::*;
